+++ EssentialApp/EssentialApp/SceneDelegate.swift
@@ -10,6 +10,10 @@ import EssentialFeediOS
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 	var window: UIWindow?
 
+	let localStoreURL = NSPersistentContainer
+		.defaultDirectoryURL()
+		.appendingPathComponent("feed-store.sqlite")
+
 	func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
 		guard let _ = (scene as? UIWindowScene) else { return }
 		
@@ -18,16 +22,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 		let remoteClient = makeRemoteClient()
 		let remoteFeedLoader = RemoteFeedLoader(url: remoteURL, client: remoteClient)
 		let remoteImageLoader = RemoteFeedImageDataLoader(client: remoteClient)
-		
-		let localStoreURL = NSPersistentContainer
-			.defaultDirectoryURL()
-			.appendingPathComponent("feed-store.sqlite")
-
-		#if DEBUG
-		if CommandLine.arguments.contains("-reset") {
-			try? FileManager.default.removeItem(at: localStoreURL)
-		}
-		#endif
 		
 		let localStore = try! CoreDataFeedStore(storeURL: localStoreURL)
 		let localFeedLoader = LocalFeedLoader(store: localStore, currentDate: Date.init)
@@ -46,26 +40,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 					cache: localImageLoader)))
 	}
 	
-	private func makeRemoteClient() -> HTTPClient {
-		#if DEBUG
-		if UserDefaults.standard.string(forKey: "connectivity") == "offline" {
-			return AlwaysFailingHTTPClient()
-		}
-		#endif
-		
+	func makeRemoteClient() -> HTTPClient {
 		return URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
 	}
 }
-
-#if DEBUG
-private class AlwaysFailingHTTPClient: HTTPClient {
-	private class Task: HTTPClientTask {
-		func cancel() {}
-	}
-	
-	func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-		completion(.failure(NSError(domain: "offline", code: 0)))
-		return Task()
-	}
-}
-#endif
